% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{Plot3D}
\alias{Plot3D}
\title{Plot a 3D scatter plot}
\usage{
Plot3D(df, col = "cornflowerblue")
}
\arguments{
\item{df}{A data frame with 3 columns representing the x, y, and z coordinates.}

\item{col}{The color of the points in the plot. Default is "cornflowerblue".}
}
\value{
None
}
\description{
This function plots a 3D scatter plot using the \code{rgl} package.
}
\examples{
df <- data.frame(A = c(1, 2, 3), B = c(4, 5, 6), C = c(7, 8, 9))
Plot3D(df)

}
